前后端基本完成，差验证和最终整合

+++ src/main/java/com/pan/auctionsystem/model/AuctionItemForRedis.java
@@ -10,8 +10,8 @@ public class AuctionItemForRedis {
     Long itemShelfDate;
     Long itemStartDate;
     Long itemEndDate;
-    double itemStartingPrice;
-    double itemIncreasePrice;
+    int itemStartingPrice;
+    int itemIncreasePrice;
     String itemPicture;
 
     int itemShopId;
